move resources to components
