+++ client/src/components/resources/CellLine/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import { SearchResult, SearchResultDetail } from '../../components/SearchResult'
-import DetailsTable from '../../components/DetailsTable'
-import { HeaderRow } from '../../components/HeaderRow'
+import { SearchResult, SearchResultDetail } from '../../SearchResult'
+import DetailsTable from '../../DetailsTable'
+import { HeaderRow } from '../../HeaderRow'
 
 // TODO:: update meta data to relect new attributes
 export const CellLineSearchResult = ({ resource }) => {
