+++ client/src/components/resources/index.js
