+++ client/src/components/resources/resources.stories.js
@@ -1,33 +1,17 @@
 import * as React from 'react'
 import { Grommet, Box } from 'grommet'
 import { storiesOf } from '@storybook/react'
-import { getReadable } from '../helpers/readableNames'
-import theme from '../theme'
-import { Mappings, SearchResult, ResourceDetails } from './index'
+import { getReadable } from '../../helpers/readableNames'
+import theme from '../../theme'
+import { Mappings, SearchResult, ResourceDetails } from '.'
 
-// import test data from api folder
-import { materials } from '../../../api/dev_data/materials.json'
+import { materialsTestData } from '../../helpers/testData'
 
 const data = {}
 
-const convertTruthies = (obj) => {
-  const keys = Object.keys(obj)
-  const convertedObj = obj
-
-  keys.forEach((key) => {
-    if (obj[key] === 'True') convertedObj[key] = true
-    if (obj[key] === 'False') convertedObj[key] = false
-    if (typeof obj[key] === 'object')
-      convertedObj[key] = convertTruthies(obj[key])
-  })
-
-  return convertedObj
-}
-
-materials.forEach((material) => {
+materialsTestData.forEach((material) => {
   data[material.category] = data[material.category] || []
-  const convertedMaterial = convertTruthies(material)
-  data[material.category].push(convertedMaterial)
+  data[material.category].push(material)
 })
 
 for (const key of Object.keys(Mappings)) {
